finwise/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.13
langchain-groq>=0.0.1
langchain-huggingface>=0.0.1
langsmith>=0.0.72
chromadb>=0.4.22
pymupdf>=1.23.8
sentence-transformers>=2.2.2
transformers>=4.37.2
flashrank>=0.1.4
torch>=2.1.2
python-dotenv>=1.0.0
streamlit>=1.31.0
tqdm>=4.66.1
groq>=0.4.0
accelerate>=0.25.0